# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Criada com o framework gin",
        "title": "Hero API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/secured/hero": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Registrar dados do herói.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hero"
                ],
                "summary": "Registrar dados do herói.",
                "parameters": [
                    {
                        "description": "hero",
                        "name": "hero",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Hero"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/secured/hero/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Buscar dados do herói pelo ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hero"
                ],
                "summary": "Buscar dados do herói pelo ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do herói",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Hero"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Atualizar dados do herói.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hero"
                ],
                "summary": "Atualizar dados do herói.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do herói",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "hero",
                        "name": "hero",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Hero"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Deletar dados do herói.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hero"
                ],
                "summary": "Deletar dados do herói.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do herói",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/secured/heros": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Consultar dados dos heróis..",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hero"
                ],
                "summary": "Consultar dados dos heróis.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Hero"
                        }
                    }
                }
            }
        },
        "/api/secured/user/register": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Cadastrar usuário.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Cadastrar usuário.",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/token/": {
            "post": {
                "description": "Gerar Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Gerar Token.",
                "parameters": [
                    {
                        "description": "token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Hero": {
            "type": "object",
            "properties": {
                "Equipe": {
                    "type": "string"
                },
                "Nome": {
                    "type": "string"
                },
                "Super Poder": {
                    "type": "string"
                }
            }
        },
        "model.TokenRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}